[core] Fix how TorrentFile's are created
Now we can handle multiple consecutive zero length files while
retaining the new logic to binary search files either by offset
within the overall torrent, or by piece index.

As part of this the test suite has been updated to ensure it
always sets the start and end piece index correctly, and also
the offset within the torrent file.

+++ src/MonoTorrent.Tests/MonoTorrent.Client.Modes/DownloadModeTests.cs
@@ -60,7 +60,7 @@ namespace MonoTorrent.Client.Modes
             ConnectionManager = new ConnectionManager ("LocalPeerId", Settings, DiskManager);
             TrackerManager = new ManualTrackerManager ();
 
-            int[] fileSizes = {
+            long[] fileSizes = {
                 Piece.BlockSize / 2,
                 Piece.BlockSize * 32,
                 Piece.BlockSize * 2,
